chore: Refactor project and user controllers and services
The code changes involve refactoring the project and user controllers and services. The `useeModel` import has been corrected to `userModel` in the `projectController.js` file. Additionally, new controller functions `getAllProjects` and `addUserToProject` have been added to the `projectController.js` file. In the `userController.js` file, a new controller function `getAllUsersController` has been added. The `projectRoutes.js` and `userRoutes.js` files have been updated to include the new routes. The `projectService.js` and `userService.js` files have been updated to include the new service functions.

This commit message follows the convention of starting with a verb in the imperative form and providing a concise summary of the changes made.